[table service][common] add BytesHashRouter
*Motivation*

Currently the hash router takes `ByteBuf`. At some cases,
we need to compute the hash of bytes array directly.

*Changes*

Add `BytesHashRouter` to compute hash value of bytes array directly,
avoiding wrapping bytes into `ByteBuf`

+++ stream/common/src/main/java/org/apache/bookkeeper/common/hash/Murmur3.java
@@ -208,7 +208,7 @@ public class Murmur3 {
      * @param seed   - seed. (default is 0)
      * @return - hashcode (2 longs)
      */
-    public static long[] hash128(byte[] data, int offset, int length, int seed) {
+    public static long[] hash128(byte[] data, int offset, int length, long seed) {
         long h1 = seed;
         long h2 = seed;
         final int nblocks = length >> 4;
